add exercises 4.15 - 4.23'

+++ controllers/blogs.js
@@ -1,22 +1,52 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user");
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body);
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
+  const body = request.body;
+
+  const user = request.user;
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id,
+  });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
-});
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() !== request.decodedToken.id) {
+      return response
+        .status(403)
+        .json({ error: "you are not authorized for this action" });
+    }
+
+    const user = request.user;
+    user.blogs = user.blogs.filter((b) => b.toString() !== request.params.id);
+    await user.save();
+
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
+);
 
 blogsRouter.put("/:id", async (request, response) => {
   const blog = {
@@ -24,6 +54,7 @@ blogsRouter.put("/:id", async (request, response) => {
     author: request.body.author,
     url: request.body.url,
     likes: request.body.likes,
+    user: request.body.user,
   };
 
   const updatedBlogs = await Blog.findByIdAndUpdate(request.params.id, blog);
